Move code from IRelativeHttpClient to IHttpClient.
